Simplify dictionary creation in sessionDetails.data
The code change modifies the way dictionaries are created within the `Details` property of the `sessionDetails.data` collection. Instead of creating a dictionary with explicit "Key" and "Value" entries, the new code directly uses the `name` and `value` properties of `detail.value` as the key-value pairs in the dictionary. This makes the code more concise and potentially easier to read and maintain.